Door jumping movement
Implemented the movement through doors by creating prefabs for the door entries and exits. Can be implemented by adding to the waveConfig. The movement itself is implemented to the enemyPathing script, modifying the move() function. A delay is yet to be added(tip pls :(). Minor adjustments to the existing path was made.