Pause Menu (squashed)
Deleted the quit and main menu buttons, having the resume one be the only one remaining, and then added the canvas to the main scene

Pause Menu Script Update

Added the serialized fields for the buttons and updated the implementation for the listeners.

Update PauseMenu.cs

Fixed the issue with having the pause menu not show up, it was in the Pause Menu animator settings, putting the pause menu update mode to unscaled time. The issue that remains is getting the button listeners to listen.

Finished

The finished paused menu has only the resume functionality.

Actual Finished one

added a }

Pause Menu screen fader

Cleaned up the hierarchy of the pauseMenu.cs and added a script fader, the two are completely implemented, just have to see how to organize them in the canvas.

Updating Scenes

Updating the scene elements to accommodate the scripts.

Polished it up (FInished)

added the fader animation for the pause menu and implemented the quit and load menu buttons. Also cleaned up the pause menu script by getting rid of the serialized button variables and attaching the buttons to the functions through the onClick option in each buttons inspector. Got rid of the fader script because of panel implementation.

Co-Authored-By: Kenny Averna <68562756+kennyav@users.noreply.github.com>