Pause Menu
* Table addition

- 2 tables were added
- location of tables were justified

* Added Table and Chair to TestMainScene

- One chair and one table in the main hall

* Waiting Chairs

Test commit to github that added three waiting chairs into the test main scene.

* Pause Menu

Created the menu option under the canvas in the Test_MainSection* scene. Started a script for the menu as well however there is a bug in getting the menu to appear after pressing the pause button.

* Revert "Pause Menu"

This reverts commit ffb408894720abdc7e15f6a2cee6838923d4a8b9.

* Re-imported PauseMenu.cs

Created by Kenny A
Added comments by Payam F

* Pause Menu (squashed)

Deleted the quit and main menu buttons, having the resume one be the only one remaining, and then added the canvas to the main scene

Pause Menu Script Update

Added the serialized fields for the buttons and updated the implementation for the listeners.

Update PauseMenu.cs

Fixed the issue with having the pause menu not show up, it was in the Pause Menu animator settings, putting the pause menu update mode to unscaled time. The issue that remains is getting the button listeners to listen.

Finished

The finished paused menu has only the resume functionality.

Actual Finished one

added a }

Pause Menu screen fader

Cleaned up the hierarchy of the pauseMenu.cs and added a script fader, the two are completely implemented, just have to see how to organize them in the canvas.

Updating Scenes

Updating the scene elements to accommodate the scripts.

Polished it up (FInished)

added the fader animation for the pause menu and implemented the quit and load menu buttons. Also cleaned up the pause menu script by getting rid of the serialized button variables and attaching the buttons to the functions through the onClick option in each buttons inspector. Got rid of the fader script because of panel implementation.

Co-Authored-By: Kenny Averna <68562756+kennyav@users.noreply.github.com>

* Settings Menus

Created a test scene for the settings menu. Started on the layout with the three sliders and back button. Created a Settings script and began implementing it.

* Cleanup Menus branch

* cleanup 2

Co-authored-by: payam <payam.4uzandeh@gmail.com>
Co-authored-by: YanNaingWin1 <ynwin1@student.ubc.ca>
Co-authored-by: Kenny Averna <kaverna@student.ubc.ca>
Co-authored-by: Kenny Averna <68562756+kennyav@users.noreply.github.com>